+++ server/path.utils.ts
@@ -1,0 +1,7 @@
+import * as fs from "fs";
+
+export function createDirIfNotExists(dir: string) {
+  if (!fs.existsSync(dir)){
+    fs.mkdirSync(dir);
+  }
+}